issue #3 resolved

+++ install.sh
@@ -1,6 +1,14 @@
-#!/bin/bash
+#!/bin/bash -e
 
 dir=~/.local/share/gnome-shell/extensions/vpn-indicator@howdoicomputer.fastmail.com
 
-rm -rf $dir
+if [ -d $dir ]; then
+  rm -rf $dir
+else
+  echo "brand-new install..."
+fi
+mkdir -p $dir
 cp -R . $dir
+
+echo "installed:"
+ls -al $dir
